@@ -4,7 +4,7 @@ version: 0.4.0
 homepage: "https://github.com/JordanADavies/liquid_progress_indicator"
 
 environment:
-  sdk: '>=2.12.0 <3.0.0'
+  sdk: '>=2.17.1 <3.0.0'
 
 dependencies:
   flutter:
