@@ -2,11 +2,13 @@ import 'package:flutter/material.dart';
 import 'package:liquid_progress_indicator/liquid_progress_indicator.dart';
 
 class LiquidCustomProgressIndicatorPage extends StatelessWidget {
+  const LiquidCustomProgressIndicatorPage({super.key});
+
   @override
   Widget build(BuildContext context) {
     return Scaffold(
       appBar: AppBar(
-        title: Text("Liquid Custom Progress Indicators"),
+        title: const Text("Liquid Custom Progress Indicators"),
       ),
       body: Column(
         mainAxisAlignment: MainAxisAlignment.spaceEvenly,
@@ -23,7 +25,7 @@ class LiquidCustomProgressIndicatorPage extends StatelessWidget {
               LiquidCustomProgressIndicator(
                 direction: Axis.horizontal,
                 backgroundColor: Colors.grey[300],
-                valueColor: AlwaysStoppedAnimation(Colors.red),
+                valueColor: const AlwaysStoppedAnimation(Colors.red),
                 shapePath: _buildSpeechBubblePath(),
               ),
             ],
@@ -69,14 +71,14 @@ class _AnimatedLiquidCustomProgressIndicator extends StatefulWidget {
 class _AnimatedLiquidCustomProgressIndicatorState
     extends State<_AnimatedLiquidCustomProgressIndicator>
     with SingleTickerProviderStateMixin {
-  AnimationController _animationController;
+ late AnimationController _animationController;
 
   @override
   void initState() {
     super.initState();
     _animationController = AnimationController(
       vsync: this,
-      duration: Duration(seconds: 10),
+      duration: const Duration(seconds: 10),
     );
 
     _animationController.addListener(() => setState(() {}));
@@ -97,11 +99,11 @@ class _AnimatedLiquidCustomProgressIndicatorState
         value: _animationController.value,
         direction: Axis.vertical,
         backgroundColor: Colors.white,
-        valueColor: AlwaysStoppedAnimation(Colors.green),
+        valueColor: const AlwaysStoppedAnimation(Colors.green),
         shapePath: _buildHeartPath(),
         center: Text(
           "${percentage.toStringAsFixed(0)}%",
-          style: TextStyle(
+          style: const TextStyle(
             color: Colors.lightGreenAccent,
             fontSize: 20.0,
             fontWeight: FontWeight.bold,
